+++ R/01-data_preprocessing.R
@@ -45,3 +45,54 @@ data_preprocessing <- function(db_list, matching_fields) {
   } )
   return(preprocessed_data)
 }
+
+
+#' Merge multiple input files from the same source
+#'
+#' @details
+#' It is fairly common to retrieve data from a single bibliographic database in small chunks. Thus, this function is designed to merge multiple files from the same source into a single file while also removing duplicate records.
+#'
+#'
+#' @param input_files - an array containing the path to all input files
+#' @param sep - field separator. Default: comma (',')
+#' @param quote - quote type used for character fields. Default: Double quotes ('"')
+#'
+#' @return a single dataframe with all unique records from the input files
+#' @export
+#'
+#' @examples
+#'
+#' ## Generating tempfiles
+#' tempfile1 <- tempfile(fileext = ".csv")
+#' tempfile2 <- tempfile(fileext = ".csv")
+#' write.csv(ufrj_bio_0122$Biochemistry, file = tempfile1, row.names = FALSE)
+#' write.csv(ufrj_bio_0122$Genetics, file = tempfile2, row.names = FALSE)
+#'
+#' ## Testing function
+#' merged_files <- merge_input_files(c(tempfile1, tempfile2))
+#' dim(merged_files)
+#' head(merged_files)
+#'
+merge_input_files <- function(input_files,
+                              sep = ",",
+                              quote = '"') {
+  df_list <- lapply(input_files, function(input_file) {
+    utils::read.csv(input_file,
+             sep = sep,
+             quote = quote,
+             strip.white = TRUE,
+             check.names = FALSE) })
+  tryCatch({
+    df <- do.call(rbind, df_list)
+  }, error = function(err) {
+    stop('Failed to merge files. Are they from the same database and/or have the same columns?') }
+  )
+  df[] <- lapply(df, function(col) { #Cleaning data (one column at a time)
+    col <- trimws(as.character(col)) # Removing leading and trailing whitespaces
+    col[which(col == "" | is.null(col))] <- NA  # Convert empty or null values to NA
+    return(col)
+  })
+  df <- df[!duplicated(df), ]   # Removing duplicate records
+
+  return( df )
+}